WT_LSM_TREE.chunk_count_limit is now a uint32_t, cast when reading the metadata entry.

+++ src/lsm/lsm_meta.c
@@ -79,7 +79,7 @@ __wt_lsm_meta_read(WT_SESSION_IMPL *session, WT_LSM_TREE *lsm_tree)
 		else if (WT_STRING_MATCH("bloom_hash_count", ck.str, ck.len))
 			lsm_tree->bloom_hash_count = (uint32_t)cv.val;
 		else if (WT_STRING_MATCH("chunk_count_limit", ck.str, ck.len)) {
-			lsm_tree->chunk_count_limit = (uint64_t)cv.val;
+			lsm_tree->chunk_count_limit = (uint32_t)cv.val;
 			if (cv.val != 0)
 				F_CLR(lsm_tree, WT_LSM_TREE_MERGES);
 		} else if (WT_STRING_MATCH("chunk_max", ck.str, ck.len))
